refactor: move `RtcStats` convertation into separate module
